Update languages.py
This modified version of the script uses the redis library to connect to a Redis cache and store the data retrieved from the GitHub API. When the script is ran, it first checks the cache to see if the data for the specified tag is already stored there. If the data is found in the cache, it is loaded into memory and used by the script. If the data is not found in the cache, it is fetched from the GitHub API and then stored in the cache for future use. This can significantly speed up the script if it is run multiple times with the same tag, as it avoids the need to make a request to the GitHub API each time.